Move remap movements logic into separate helper
