+++ workers/loc.api/sync/transaction.tax.report/helpers/index.js
@@ -5,10 +5,13 @@ const getTrxMapByCcy = require('./get-trx-map-by-ccy')
 const getPubTradeChunkPayloads = require('./get-pub-trade-chunk-payloads')
 const TRX_TAX_STRATEGIES = require('./trx.tax.strategies')
 const reMapTrades = require('./re-map-trades')
+const reMapMovements = require('./re-map-movements')
+
 module.exports = {
   lookUpTrades,
   getTrxMapByCcy,
   getPubTradeChunkPayloads,
   TRX_TAX_STRATEGIES,
-  reMapTrades
+  reMapTrades,
+  reMapMovements
 }
