+++ workers/loc.api/sync/transaction.tax.report/index.js
@@ -1,9 +1,6 @@
 'use strict'
 
 const {
-  isForexSymb
-} = require('../helpers')
-const {
   TrxTaxReportGenerationTimeoutError
 } = require('../../errors')
 const {
@@ -11,7 +8,8 @@ const {
   getTrxMapByCcy,
   getPubTradeChunkPayloads,
   TRX_TAX_STRATEGIES,
-  reMapTrades
+  reMapTrades,
+  reMapMovements
 } = require('./helpers')
 
 const { decorateInjectable } = require('../../di/utils')
@@ -208,40 +206,10 @@ class TransactionTaxReport {
       trades,
       { remapedTrxs, remapedTrxsForConvToUsd }
     )
-
-    for (const movement of movements) {
-      if (
-        !movement?.currency ||
-        isForexSymb(movement.currency) ||
-        !Number.isFinite(movement?.amount) ||
-        movement.amount === 0 ||
-        !Number.isFinite(movement?.mtsUpdated)
-      ) {
-        continue
-      }
-
-      const firstSymb = movement.currency
-      const lastSymb = 'USD'
-      const symbSeparator = firstSymb.length > 3
-        ? ':'
-        : ''
-
-      const remapedMovement = {
-        isMovements: true,
-        symbol: `t${firstSymb}${symbSeparator}${lastSymb}`,
-        mtsCreate: movement.mtsUpdated,
-        firstSymb,
-        lastSymb,
-        firstSymbPrice: null,
-        lastSymbPrice: 1,
-        execAmount: movement.amount,
-        // NOTE: execPrice = firstSymbPrice and should be set when converting currencies
-        execPrice: 0
-      }
-
-      remapedTrxs.push(remapedMovement)
-      remapedTrxsForConvToUsd.push(remapedMovement)
-    }
+    reMapMovements(
+      movements,
+      { remapedTrxs, remapedTrxsForConvToUsd }
+    )
 
     const trxs = remapedTrxs
       .sort((a, b) => b?.mtsCreate - a?.mtsCreate)
